tests/wpt/tests/css/css-borders/tentative/parsing/corner-shape-computed.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Borders and Box Decorations 4 Test: Computed values of 'corner-shape'</title>
<link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#corner-shaping">
<meta name="assert" content="This test checks that the computed value of 'corner-shape' is correct.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>

<div id="target"></div>

<script>
test_computed_value("corner-top-left-shape", "round");
test_computed_value("corner-top-left-shape", "scoop");
test_computed_value("corner-top-left-shape", "superellipse(5)");
test_computed_value("corner-top-left-shape", "superellipse(0.2)");
test_computed_value("corner-top-left-shape", "superellipse(0)", "notch");
test_computed_value("corner-top-left-shape", "superellipse(infinity)", "straight");
test_computed_value("corner-top-left-shape", "superellipse(2)", "round");
test_computed_value("corner-top-left-shape", "superellipse(1)", "bevel");
test_computed_value("corner-top-left-shape", "superellipse(4)", "squircle");
test_computed_value("corner-top-left-shape", "superellipse( .5)", "scoop");
test_computed_value("corner-top-right-shape", "round");
test_computed_value("corner-top-right-shape", "superellipse(5)");
test_computed_value("corner-bottom-right-shape", "scoop");
test_computed_value("corner-bottom-left-shape", "superellipse(5)");
</script>
